requirements.txt
groq>=0.13.0
streamlit>=1.42.0
python-dotenv>=1.0.0
chromadb>=0.5.0
pyyaml>=6.0
